Added example v1beta1 test component
