--- api-ctk/TMF637-ProductInventory/Windows-Bat-RUNCTK.bat
@@ -1,4 +1,0 @@
-@echo off
-cd .\ctk
-npm install && npm start
-cmd
