--- api-ctk/TMF637-ProductInventory/config.json
@@ -1,8 +1,0 @@
-{
-    "url": " http://localhost:8637/tmf-api/productInventory/v4/",
-    "headers": {
-        "Accept": "application/json",
-        "Content-Type": "application/json"
-    },
-    "payloads": {}
-}
