--- api-ctk/TMF637-ProductInventory/ctk/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "tmf637-productinventory",
-  "version": "4.0.0",
-  "description": "Conformance Test Kit for TMF637-ProductInventory",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
-  },
-  "keywords": [
-    "TMForum",
-    "Conformance",
-    "Test",
-    "Kit",
-    "CTK",
-    "TMF637-ProductInventory"
-  ],
-  "author": "Henrique Rodrigues (hrodrigues@tmforum.org)",
-  "dependencies": {
-    "ajv": "^6.12.3",
-    "jsonschema": "^1.2.6",
-    "newman": "^4.0.2",
-    "newman-reporter-html": "^1.0.1",
-    "readline": "^1.3.0",
-    "request": "^2.88.0",
-    "request-promise": "^4.2.2"
-  }
-}
