+++ api-ctk/TMF637-ProductInventory_V4-0-0/LICENSE
@@ -1,0 +1,13 @@
+RAND licence mode: Notice
+Copyright © TM Forum 2020. All Rights Reserved.
+This document and translations of it may be copied and furnished to others, and derivative works that comment on or otherwise explain it or assist in its implementation may be prepared, copied, published, and distributed, in whole or in part, without restriction of any kind, provided that the above copyright notice and this section are included on all such copies and derivative works. However, this document itself may not be modified in any way, including by removing the copyright notice or references to TM FORUM, except as needed for the purpose of developing any document or deliverable produced by a TM FORUM Collaboration Project Team (in which case the rules applicable to copyrights, as set forth in the TM FORUM IPR Policy must be followed) or as required to translate it into languages other than English.
+The limited permissions granted above are perpetual and will not be revoked by TM FORUM or its successors or assigns.
+This document and the information contained herein is provided on an "AS IS" basis and TM FORUM DISCLAIMS ALL WARRANTIES, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO ANY WARRANTY THAT THE USE OF THE INFORMATION HEREIN WILL NOT INFRINGE ANY OWNERSHIP RIGHTS OR ANY IMPLIED WARRANTIES OF MERCHANTABILITY OR FITNESS FOR A PARTICULAR PURPOSE.
+Direct inquiries to the TM Forum office:
+4 Century Drive, Suite 100
+
+
+Parsippany, NJ 07054USA
+Tel No. +1 973 944 5100
+Fax No. +1 973 944 5110
+TM Forum Web Page: www.tmforum.org
