+++ api-ctk/TMF637-ProductInventory_V4-0-0/Windows-Bat-RUNCTK.bat
@@ -1,0 +1,4 @@
+@echo off
+cd .\ctk
+npm install && npm start
+cmd
