+++ api-ctk/TMF637-ProductInventory_V4-0-0/ctk/TMFENV-V4.0.0.postman_environment.json
@@ -1,0 +1,14 @@
+{
+	"id": "da32588a-3916-4d78-a546-7623e391f650",
+	"name": "TMFENV-V4.0.0",
+	"values": [
+        {
+			"key": "Product_Inventory",
+			"value": "http://localhost/tmf-api/",
+			"enabled": true
+		}
+	],
+	"_postman_variable_scope": "environment",
+	"_postman_exported_at": "2020-07-07T14:17:10.805Z",
+	"_postman_exported_using": "Postman/7.27.1"
+}
