+++ package.json
@@ -4,12 +4,12 @@
   "description": "ODA Component Compliance Test Kit CTK",
   "main": "index.js",
   "scripts": {
-    "start": "node L1-staticValidationRunner.js",
-    "L1-static": "node L1-staticValidationRunner.js",
-    "L1-dynamic": "node L1-dynamicValidationRunner.js",
-    "L2-static": "node L2-staticValidationRunner.js",
-    "L2-dynamic": "node L2-dynamicValidationRunner.js",
-    "test": "node L1-staticValidationRunner.js"
+    "start": "node L1-static-CTK.js",
+    "L1-static": "node L1-static-CTK.js",
+    "L1-dynamic": "node L1-dynamic-CTK.js",
+    "L2-static": "node L2-static-CTK.js",
+    "L2-dynamic": "node L2-dynamic-CTK.js",
+    "test": "node L1-static-CTK.js"
   },
   "repository": {
     "type": "git",
